pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "pyhabot"
version = "0.1.0"
description = "Async Python bot that monitors HardverApró (Hungarian classified ads site) search result pages"
authors = ["radamhu <radamhu@example.com>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/radamhu/PYHABOT"
repository = "https://github.com/radamhu/PYHABOT"
keywords = ["bot", "scraping", "discord", "telegram", "hardverapro", "monitoring"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
packages = [{include = "pyhabot", from = "src"}]

[tool.poetry.scripts]
pyhabot = "pyhabot.cli:main"

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "2.5.0"
python-dotenv = "1.0.1"
tinydb = "4.8.2"
beautifulsoup4 = "4.13.3"
aiohttp = "^3.9.0"
telegram-py = {git = "https://github.com/ilovetocode2019/telegram.py", rev = "1c09546"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
mypy = "^1.5.0"
ruff = "^0.1.0"
black = {version = "^23.7.0", optional = true}
pre-commit = {version = "^3.3.0", optional = true}

[tool.poetry.extras]
dev = ["black", "pre-commit"]

[tool.ruff]
line-length = 88
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "N",  # pep8-naming
    "D",  # pydocstyle
    "S",  # flake8-bandit
    "A",  # flake8-builtins
    "T20", # flake8-print
    "PT", # flake8-pytest-style
    "RET", # flake8-return
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
    "B904",  # Allow raising exceptions without from e, for HTTPException
    "D100", # Missing docstring in public module
    "D101", # Missing docstring in public class
    "D102", # Missing docstring in public method
    "D103", # Missing docstring in public function
    "D104", # Missing docstring in public package
    "D105", # Missing docstring in magic method
    "D107", # Missing docstring in __init__
    "D203", # 1 blank line required before class docstring
    "D213", # Multi-line docstring summary should start at the second line
    "S101", # Use of assert detected
    "T201", # print found
    "PT001", # Use @pytest.fixture without parentheses
    "PT023", # Use @pytest.mark.anyio instead of @pytest.mark.trio
]

[tool.ruff.per-file-ignores]
"tests/*" = [
    "E501",  # line too long in tests
    "S101",  # assert is fine in tests
    "S105",  # hardcoded passwords in tests are fine
    "S106",  # hardcoded passwords in tests are fine
    "S108",  # hardcoded temporary file in tests are fine
    "D",  # docstrings not required in tests
    "PT",  # some pytest style rules are too strict
]
"src/pyhabot/adapters/integrations/terminal.py" = ["T201"]  # print is fine in terminal integration

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_equality = true
show_error_codes = true
show_column_numbers = true
show_error_context = true
pretty = true
warn_no_return = true
warn_unreachable = true
strict_optional = true

[[tool.mypy.overrides]]
module = [
    "telegram_py.*",
    "telegrampy.*",
    "discord.*",
    "tinydb.*",
    "beautifulsoup4.*",
    "aiohttp.*",
    "pytest.*",
    "pytest_asyncio.*",
    "pytest_cov.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
check_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=pyhabot",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]